Handle windows explicitly.
I am not sure this is the way I really want to handle the problem. It
is clear that the current window is the wrong thing to use when
sending off packets to the other end, so there is some window-handling
to be done in order to figure out what is going on here.