Drop packets under enobufs.
If we get {error, enobufs} the send buffer is exhausted. This may
happen, so what we do is to log a warning, but otherwise ignore the
thing as if the packet were dropped. The reason we can do this is
because we gamble that a later ACK packet will actually ACK for this
one as well, so it doesn't really matter that much.