Detect FIN packets with no payload.
It is the code updating the receive buffer which tracks the got_fin
message back to the sender. If the packet has no payload, this
analysis was skipped and we simply just returned ok. This update
changes the code so it correctly detects the receipt of a got_fin
message.