Fixed a duplicate incoming SYN packet problem.
The problem was that when we get duplicate SYN packets, these are not
blindly forwarded since they come in with a ConnID one too low. This
means we look up the wrong registry entry and thus we create two
processes. This of course messes a lot of internal state up and we are
then thrown into a hell of maintenance with no hope of ever coming up
again to the real world.

The fix is rather easy. If the packet incoming is a SYN packet, we
have to check it against CID+1 and not CID. This means that the
duplicate SYN packets will not enter the accept queue but be forwarded
to the right receiver.