Solve a problem: ledbat timer order.
It looked very much like we had a bug where the ledbat timer was set
and then triggered right away. But really, the thing that happened was
that we set another timer just before we handled the currently
triggered timer. So it was a fluke.

While here, go through the TODO.org list yet again and update with
more information where solutions have become known.