Quickly throw out a eclosed to senders.
This is actually a design decision by me. If we know that we are in
GOT_FIN, the connection is for all purposes in a closed state. We may
have numerous senders awaiting to send data, but they will never ever
complete in fact. And we know it. So we just push up an {error,
eclosed} so the application process knows what has happened.